requirements.txt
langchain-text-splitters>=1.1.1
pydantic-settings>=2.13.1
pypdf>=6.7.5
faiss-cpu>=1.13.2
fastapi>=0.133.0
langchain>=1.2.10
langgraph>=1.0.9
pydantic>=2.12.5
pydantic-settings>=2.13.1
sentence-transformers>=5.2.3
uvicorn>=0.41.0
accelerate>=1.12.0
bitsandbytes>=0.46.1